Improvement of Bot AI
many many update, refer to @fradunz 's core, thank you!

+++ src/game/Objects/CreatureDefines.h
@@ -289,7 +289,8 @@ struct CreatureData
     float wander_distance = 0.0f;
     float health_percent = 100.0f;
     float mana_percent = 100.0f;
-    uint8 movement_type = 0;
+    //uint8 movement_type = 0;
+	uint32 movement_type = 0;
     uint32 spawn_flags = 0;
     float visibility_mod = 0.0f;
